requirements_fixed.txt
absl-py
appdirs
asgiref
astunparse
audioread
boto3
botocore
cachetools
certifi
cffi
charset-normalizer
crispy-bootstrap5
cycler
decorator
dj-database-url
Django
django-crispy-forms
django-heroku
flatbuffers
fonttools
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
gunicorn
h5py
idna
importlib-metadata
jmespath
joblib
keras
Keras-Preprocessing
kiwisolver
libclang
llvmlite
Markdown
matplotlib
numba
numpy
oauthlib
opencv-python
opt-einsum
packaging
pandas
Pillow
pooch
protobuf
psycopg2-binary
pyasn1
pyasn1-modules
pycparser
pyparsing
python-dateutil
python-decouple
pytz
requests
requests-oauthlib
resampy
rsa
s3transfer
scikit-learn
scipy
six
sqlparse
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
termcolor
threadpoolctl
torch
torchvision
typing_extensions
tzdata
urllib3
Werkzeug
whitenoise
wrapt
zipp
